requirements.txt
streamlit==1.32.0
PyPDF2==3.0.1
python-docx==1.1.0
openai==1.12.0
boto3==1.34.69
botocore==1.34.69
python-dotenv==1.0.1
pytest==7.3.1
